refactor(ngcc): move analyze and compile functions into their own files
